単一ファイルコンポーネント
