+++ src/components/forms/InputText.vue
@@ -1,0 +1,24 @@
+<template>
+  <input class="input" v-model="sharedState.state.property.nickname" placeholder="NICKNAMEを入力してください。">
+</template>
+
+<script>
+//  インポートします。
+import PropertyStore from '../../models/PropertyStore.js'
+
+export default {
+  name: 'nickname_input',
+  data: function () {
+    return {
+      privateState: {},
+      sharedState: PropertyStore
+    }
+  }
+}
+</script>
+<style scoped>
+.input{
+  width: 130pt;
+  height:30pt;
+}
+</style>
